+++ resources/assets/js/components/Pagebuilder/ModuleSection.vue
@@ -1,18 +1,20 @@
 <template>
-    <section @mouseenter="mouseEnter" @mouseleave="mouseLeave" class="{section.id} section-item stack block inset relative">
-        <h3 class="pagebuilder-section-title" v-if="title" v-text="title"></h3>
+    <section @mouseenter="mouseEnter" @mouseleave="mouseLeave" class="{section.id} border-l-2 bg-grey-100 border-primary-500 stack block inset relative rounded-r">
+        
+        <h3 class="text-primary-500 mb-0 pb-2 font-normal" v-if="title" v-text="title"></h3>
+
         <div class="row to-minimize">
             <div class="column-6">
                 <chief-multiselect
-                        :name="'sections['+sectionKey+']['+_uid+']'"
-                        :options='options'
-                        :multiple="false"
-                        :selected="this.selected"
-                        grouplabel="group"
-                        groupvalues="values"
-                        labelkey="label"
-                        valuekey="id"
-                        :placeholder="placeholder"
+                    :name="'sections['+sectionKey+']['+_uid+']'"
+                    :options='options'
+                    :multiple="false"
+                    :selected="this.selected"
+                    grouplabel="group"
+                    groupvalues="values"
+                    labelkey="label"
+                    valuekey="id"
+                    :placeholder="placeholder"
                 >
                 </chief-multiselect>
             </div>
@@ -21,11 +23,15 @@
         <pagebuilder-menu :section="section"></pagebuilder-menu>
 
         <div class="module-icons-left">
-            <span class="grip-button icon icon-menu inset-xs"></span>
+            <span class="grip-button inset-xs flex justify-center text-grey-500 text-center my-2 cursor-move">
+                <svg width="18" height="18"><use xlink:href="#menu"/></svg>
+            </span>
         </div>
 
         <div class="module-icons-right">
-            <span class="delete-button icon icon-trash inset-xs" @click="removeThisSection(section.sort)"></span>
+            <span class="delete-button inset-xs flex justify-center text-error text-center my-2 cursor-pointer" @click="removeThisSection(section.sort)">
+                <svg width="18" height="18"><use xlink:href="#trash"/></svg>
+            </span>
         </div>
 
     </section>
@@ -92,18 +98,6 @@
     }
 </script>
 <style scoped>
-.delete-button {
-    color:red;
-    /* border-left: 2px solid red; */
-    margin: 0.5rem 0;
-    text-align: center;
-}
-.grip-button {
-    color: rgb(30,30,30);
-    /* border-right: 2px solid rgb(30,30,30); */
-    margin: 0.5rem 0;
-    text-align: center;
-}
 .module-icons-left {
     position: absolute;
     top: 0;
@@ -128,17 +122,14 @@
     width: 40px;
     transition: 0.15s all ease-in;
 }   
-section {
-    position: relative;
-}
 .reveal-left {
     opacity: 1;
-    left: -43px;
+    left: -42px;
     transition: 0.15s all ease-in;
 }
 .reveal-right {
     opacity: 1;
-    right: -40px;
+    right: -42px;
     transition: 0.15s all ease-in;
 }
 </style>