+++ resources/views/back/system/settings.blade.php
@@ -11,7 +11,6 @@
     <form action="{{ route('chief.back.settings.update') }}" id="updateForm" method="POST" role="form">
         {{ csrf_field() }}
         <input name="_method" type="hidden" value="PUT">
-
         @foreach($manager->fields() as $field)
             {!! $manager->renderField($field) !!}
         @endforeach