+++ tests/fakes/ArticlePageFake.php
@@ -8,6 +8,8 @@ use Thinktomorrow\Chief\Pages\Page;
 
 class ArticlePageFake extends Page
 {
+    protected static $managedModelKey = 'articles_fake';
+
     public function url($locale = null): string
     {
         return route('articles.show', 'foobar');