+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All Notable changes to the `chief` application template will be documented in th
 - Removed: Homepage setting and `chief-settings.homepage` config value
 - Removed: `ActsAsChild::viewKey()` contract method requirement. This is now the responsibility of the `ViewableContract`.
 - Removed: `ActsAsParent::viewKey()` contract method requirement. This is now the responsibility of the `ViewableContract`.
+- Removed: Honeypot middleware and helper.
 - Changed: By default the chief route `pages.show` is autoloaded by the package. This can be opted out by setting the `chief.route.autoload` value to false.
 - Added: `ViewableContract` to identify models that should be rendered on the site.
 - Changed: `ProvidesUrl` contract to identify models that should be retrievable by direct url.
@@ -16,7 +17,9 @@ All Notable changes to the `chief` application template will be documented in th
 - Added: when archiving a page another page can now be set as redirect
 - Added: improved field validation design
 - Added: `Manager::fieldsWithAssistantFields()` which also include any assistant fields. This is mostly used internally.
-
+- Removed: Homepage setting and `chief-settings.homepage` config value
+- Changed: By default the chief route `pages.show` is autoloaded by the package. This can be opted out by setting the `chief.routes.autoload` value to false.
+- Changed: don't show tabs when there is just one language in menubuilder. Case: "Of kies een eigen link."
 
 ## 0.2.14 - 2019-6-18
 - Added: add seo_image to page seo tab